Refactor import paths in quantum simulation circuit execution files to correct module structure. Update mini-app architecture diagram image file.
